// src/app/Models/theatermaster.ts
export class TheaterMasterr {
  ID: number = 0;
  CITY_ID: number = 0;
  NAME: string = '';
  CITY_NAME: string = '';
  THEATER_NAME: string = '';
  ADDRESS: string = '';
  LONGITUDE: string = '';
  LATITUDE: string = '';
  SEQUENCE_NUMBER: number = 0;
  STATUS: boolean = true;
  THEATRE_ID!: number;
  RATE_ARRAY: any;
  BOOKING_STATUS!: string;
  ACTUAL_LAYOUT_IMAGE: string='';
}